backend/requirements.txt
amqp==5.0.6
asgiref==3.4.1
autopep8==1.5.7
billiard==3.6.4.0
celery==5.0.5
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.8
django-ckeditor==6.1.0
django-cors-headers==3.10.0
django-environ==0.8.1
django-filter==21.1
django-js-asset==1.2.2
djangorestframework==3.12.4
dnspython==2.1.0
et-xmlfile==1.1.0
greenlet==1.1.2
gunicorn==20.1.0
humanize==3.12.0
kombu==5.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
Pillow==8.4.0
prometheus-client==0.11.0
prompt-toolkit==3.0.21
psycopg2==2.9.1
pycodestyle==2.8.0
pytz==2021.3
PyYAML==6.0
redis==3.5.3
six==1.16.0
sqlparse==0.4.2
tablib==3.1.0
toml==0.10.2
tornado==6.1
vine==5.0.0
wcwidth==0.2.5
xlrd==2.0.1
xlwt==1.3.0
